Move router config parsing to RouterUtil
